:zap: improve datasource informations (close #83)

+++ wlts/controller.py
@@ -87,7 +87,7 @@ class WLTS:
             describe["collection_type"] = collection.collection_type()
             describe["resolution_unit"] = {
                 "unit": collection.get_resolution_unit(),
-                "value": float(collection.get_resolution_value())
+                "value": int(collection.get_resolution_value())
             }
             describe["period"] = {
                 "start_date": collection.get_start_date(),
